refactor(home): extract HomeViewModel load/UI logic under ≤1000 LOC
Move Home UI/data models and pure loadData assembly helpers into dedicated
files, keep StateFlow ownership on HomeViewModel, and add unit tests for
the new logic.
